pyproject.toml
[project]
name = "langchainproject-new-langchainproject-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.1",
    "gradio>=5.49.1",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "langgraph>=1.0.1",
    "langserve>=0.3.3",
    "llama-index>=0.10.9",
    "pandas>=2.3.3",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
    "streamlit-chat>=0.1.1",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
